+++ src/DataManager/Program.cs
@@ -9,26 +9,27 @@ using Npgsql;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// var connStringBuilder = new NpgsqlConnectionStringBuilder {
-//   SslMode = SslMode.VerifyFull,
-//   Host = builder.Configuration["Db:Host"],
-//   Port = builder.Configuration["Db:Port"].ParseInt(),
-//   Database = builder.Configuration["Db:Name"],
-//   Username = builder.Configuration["Db:User"],
-//   Password = builder.Configuration["Db:Pass"]
-// };
+var connStringBuilder = new NpgsqlConnectionStringBuilder {
+  SslMode = SslMode.VerifyFull,
+  Host = builder.Configuration["Db:Host"],
+  Port = builder.Configuration["Db:Port"].ParseInt(),
+  Database = builder.Configuration["Db:Name"],
+  Username = builder.Configuration["Db:User"],
+  Password = builder.Configuration["Db:Pass"]
+};
 
-// var dataSourceBuilder = new NpgsqlDataSourceBuilder(connStringBuilder.ConnectionString);
-// dataSourceBuilder.EnableDynamicJson();
-// var dataSource = dataSourceBuilder.Build();
+var dataSourceBuilder = new NpgsqlDataSourceBuilder(connStringBuilder.ConnectionString);
+dataSourceBuilder.EnableDynamicJson();
+var dataSource = dataSourceBuilder.Build();
 
 // Add services to the container.
 var services = builder.Services;
 
 services
-    // .AddDbContext<DataManagerDbContext>(options => {
-    //   options.UseNpgsql(dataSource).UseSnakeCaseNamingConvention();
-    // })
+    .AddDbContext<DataManagerDbContext>(options => {
+      options.UseNpgsql(dataSource).UseSnakeCaseNamingConvention();
+    })
+    .AddScoped<ICampaignStore, CampaignStore>()
     .AddScoped<CampaignService>()
     .AddScoped<AuthService>()
     .AddHostedService<SimulationService>() // Background simulator worker (runs forever)
@@ -41,7 +42,7 @@ services.AddRazorPages();
 
 AppContext.SetSwitch("Npgsql.DisableDateTimeInfinityConversions", true);
 
-// StaticWebAssetsLoader.UseStaticWebAssets(builder.Environment, builder.Configuration); 
+StaticWebAssetsLoader.UseStaticWebAssets(builder.Environment, builder.Configuration); 
 
 var app = builder.Build();