+++ src/processor/settings/job_settings.py
@@ -4,9 +4,9 @@ from typing import Optional
 
 @dataclass
 class JobSettings:
+  models: str
   connector: Optional[str] = 'spark'
   product_id: Optional[str] = ''
   event_date: Optional[str] = ''
   action: Optional[str] = ''
-  models: Optional[str] = ''
   config_dir: Optional[str] = path.dirname(__file__) + '/../../../products'