+++ src/processor/settings/product_settings.py
@@ -3,7 +3,6 @@ from dataclasses import dataclass
 
 @dataclass
 class ProductSettings:
-  inputs: dict[str, Any]
-  datastore: dict[str, Any]
-  outputs: dict[str, Any]
+  sources: dict[str, Any]
+  targets: dict[str, Any]
   vars: dict[str, Any]