+++ src/processor/spark-run.sh
@@ -8,4 +8,4 @@ models=$4
 # SPARK_LOCAL_HOSTNAME=localhost spark-submit \
 #   --name "minerva::$product_id::$event_date::$action::$models" \
 #   --master local[*] \
-python3 src/app.py product_id=$product_id event_date=$event_date action=$action models=$models config_dir=../../products
+python3 app.py product_id=$product_id event_date=$event_date action=$action models=$models config_dir=../../products