+++ src/DataManager/CHANGELOG.md
@@ -7,7 +7,8 @@
 - Build client app with bootstrap, react, rosie-ui
 - Add OAuth2
 - Add `Simulation` tab
-  - Add `Ads Manager` screen
+  - Add `Ads Manager` screen to manage campaign
+- Auto generate campaign and save to database
 
 ## v0.0.1
 > 2025-12-07